Move syntax-related modules into Unison.Syntax
