+++ parser-typechecker/src/Unison/Result.hs
@@ -14,8 +14,8 @@ import Control.Monad.Writer
   )
 import Unison.Name (Name)
 import qualified Unison.Names.ResolutionResult as Names
-import qualified Unison.Parser as Parser
 import Unison.Prelude
+import qualified Unison.Syntax.Parser as Parser
 import Unison.Term (Term)
 import qualified Unison.Typechecker.Context as Context
