+++ parser-typechecker/src/Unison/Runtime/Debug.hs
@@ -12,8 +12,8 @@ import Debug.Trace
 import Unison.PrettyPrintEnv (PrettyPrintEnv)
 import Unison.Runtime.ANF
 import Unison.Runtime.MCode
+import Unison.Syntax.Printer.Term (pretty)
 import qualified Unison.Term as Tm
-import Unison.TermPrinter (pretty)
 import Unison.Util.EnumContainers
 import Unison.Util.Pretty (toANSI)
 import Unison.Var (Var)
