+++ parser-typechecker/src/Unison/Syntax/Parser/Type.hs
@@ -1,18 +1,18 @@
 {-# LANGUAGE OverloadedStrings #-}
 
-module Unison.TypeParser where
+module Unison.Syntax.Parser.Type where
 
 import Control.Monad.Reader (asks)
 import qualified Data.Set as Set
 import qualified Text.Megaparsec as P
 import qualified Unison.Builtin.Decls as DD
 import qualified Unison.HashQualified as HQ
-import qualified Unison.Lexer as L
 import qualified Unison.Name as Name
 import qualified Unison.NamesWithHistory as Names
-import Unison.Parser
 import Unison.Parser.Ann (Ann (..))
 import Unison.Prelude
+import qualified Unison.Syntax.Lexer as L
+import Unison.Syntax.Parser
 import Unison.Type (Type)
 import qualified Unison.Type as Type
 import Unison.Var (Var)
