+++ parser-typechecker/src/Unison/Syntax/Parsers.hs
@@ -1,10 +1,9 @@
-module Unison.Parsers where
+module Unison.Syntax.Parsers where
 
 import qualified Data.Text as Text
 import qualified Unison.Builtin as Builtin
 import qualified Unison.FileParser as FileParser
 import qualified Unison.NamesWithHistory as Names
-import qualified Unison.Parser as Parser
 import Unison.Parser.Ann (Ann)
 import Unison.Prelude
 import Unison.PrintError
@@ -12,10 +11,11 @@ import Unison.PrintError
     prettyParseError,
   )
 import Unison.Symbol (Symbol)
+import qualified Unison.Syntax.Parser as Parser
+import qualified Unison.Syntax.Parser.Term as TermParser
+import qualified Unison.Syntax.Parser.Type as TypeParser
 import Unison.Term (Term)
-import qualified Unison.TermParser as TermParser
 import Unison.Type (Type)
-import qualified Unison.TypeParser as TypeParser
 import Unison.UnisonFile (UnisonFile)
 import qualified Unison.Util.Pretty as Pr
 import Unison.Var (Var)
