+++ parser-typechecker/src/Unison/Syntax/Printer/Decl.hs
@@ -1,7 +1,7 @@
 {-# LANGUAGE OverloadedStrings #-}
 {-# OPTIONS_GHC -Wno-orphans #-}
 
-module Unison.DeclPrinter (prettyDecl, prettyDeclHeader, prettyDeclOrBuiltinHeader) where
+module Unison.Syntax.Printer.Decl (prettyDecl, prettyDeclHeader, prettyDeclOrBuiltinHeader) where
 
 import Data.List (isPrefixOf)
 import qualified Data.Map as Map
@@ -18,7 +18,6 @@ import qualified Unison.HashQualified as HQ
 import qualified Unison.Hashing.V2.Convert as Hashing
 import Unison.Name (Name)
 import qualified Unison.Name as Name
-import Unison.NamePrinter (styleHashQualified'')
 import Unison.Prelude
 import Unison.PrettyPrintEnv (PrettyPrintEnv)
 import qualified Unison.PrettyPrintEnv as PPE
@@ -26,9 +25,10 @@ import Unison.PrettyPrintEnvDecl (PrettyPrintEnvDecl (..))
 import Unison.Reference (Reference (DerivedId))
 import qualified Unison.Referent as Referent
 import qualified Unison.Result as Result
+import Unison.Syntax.Printer.Name (styleHashQualified'')
+import qualified Unison.Syntax.Printer.Type as TypePrinter
 import qualified Unison.Term as Term
 import qualified Unison.Type as Type
-import qualified Unison.TypePrinter as TypePrinter
 import qualified Unison.Typechecker as Typechecker
 import Unison.Typechecker.TypeLookup (TypeLookup (TypeLookup))
 import qualified Unison.Typechecker.TypeLookup as TypeLookup
