+++ parser-typechecker/src/Unison/Syntax/Printer/Type.hs
@@ -1,7 +1,7 @@
 {-# LANGUAGE OverloadedStrings #-}
 {-# LANGUAGE PatternSynonyms #-}
 
-module Unison.TypePrinter
+module Unison.Syntax.Printer.Type
   ( pretty,
     pretty0,
     prettyRaw,
@@ -19,13 +19,13 @@ import qualified Data.Map as Map
 import qualified Unison.Builtin.Decls as DD
 import Unison.HashQualified (HashQualified)
 import Unison.Name (Name)
-import Unison.NamePrinter (styleHashQualified'')
 import Unison.Prelude
 import Unison.PrettyPrintEnv (PrettyPrintEnv)
 import qualified Unison.PrettyPrintEnv as PrettyPrintEnv
 import Unison.PrettyPrintEnv.FQN (Imports, elideFQN)
 import Unison.Reference (Reference, pattern Builtin)
 import Unison.Referent (Referent)
+import Unison.Syntax.Printer.Name (styleHashQualified'')
 import Unison.Type
 import Unison.Util.ColorText (toPlain)
 import Unison.Util.Pretty (ColorText, Pretty, Width)
