+++ parser-typechecker/src/Unison/Typechecker/Context.hs
@@ -77,9 +77,9 @@ import qualified Unison.Pattern as Pattern
 import Unison.Prelude
 import Unison.Reference (Reference)
 import Unison.Referent (Referent)
+import qualified Unison.Syntax.Printer.Type as TP
 import qualified Unison.Term as Term
 import qualified Unison.Type as Type
-import qualified Unison.TypePrinter as TP
 import Unison.Typechecker.Components (minimize')
 import qualified Unison.Typechecker.TypeLookup as TL
 import qualified Unison.Typechecker.TypeVar as TypeVar
