+++ parser-typechecker/tests/Unison/Test/FileParser.hs
@@ -8,10 +8,10 @@ import Data.Set (elems)
 import EasyTest
 import qualified Text.Megaparsec.Error as MPE
 import Unison.FileParser (file)
-import qualified Unison.Parser as P
 import qualified Unison.Parser.Ann as P
-import Unison.Parsers (unsafeGetRightFrom, unsafeParseFileBuiltinsOnly)
 import Unison.Symbol (Symbol)
+import qualified Unison.Syntax.Parser as P
+import Unison.Syntax.Parsers (unsafeGetRightFrom, unsafeParseFileBuiltinsOnly)
 import qualified Unison.Test.Common as Common
 import Unison.UnisonFile (UnisonFile)
 import Unison.Var (Var)
