+++ unison-cli/src/Unison/Codebase/Editor/HandleCommand.hs
@@ -33,10 +33,8 @@ import qualified Unison.CommandLine.FuzzySelect as Fuzzy
 import Unison.FileParsers (parseAndSynthesizeFile, synthesizeFile')
 import qualified Unison.Hashing.V2.Convert as Hashing
 import qualified Unison.NamesWithHistory as NamesWithHistory
-import qualified Unison.Parser as Parser
 import Unison.Parser.Ann (Ann)
 import qualified Unison.Parser.Ann as Ann
-import qualified Unison.Parsers as Parsers
 import Unison.Prelude
 import qualified Unison.PrettyPrintEnv as PPE
 import qualified Unison.Reference as Reference
@@ -45,6 +43,8 @@ import qualified Unison.Server.Backend as Backend
 import qualified Unison.Server.CodebaseServer as Server
 import qualified Unison.Sqlite as Sqlite
 import Unison.Symbol (Symbol)
+import qualified Unison.Syntax.Parser as Parser
+import qualified Unison.Syntax.Parsers as Parsers
 import Unison.Term (Term)
 import qualified Unison.Term as Term
 import Unison.Type (Type)
