+++ unison-cli/src/Unison/Codebase/Editor/SlurpResult.hs
@@ -9,13 +9,13 @@ import qualified Data.Map as Map
 import qualified Data.Set as Set
 import Unison.Codebase.Editor.SlurpComponent (SlurpComponent (..))
 import qualified Unison.Codebase.Editor.SlurpComponent as SC
-import qualified Unison.DeclPrinter as DeclPrinter
 import qualified Unison.HashQualified as HQ
 import Unison.Name (Name)
 import Unison.Parser.Ann (Ann)
 import Unison.Prelude
 import qualified Unison.PrettyPrintEnv as PPE
-import qualified Unison.TypePrinter as TP
+import qualified Unison.Syntax.Printer.Decl as DeclPrinter
+import qualified Unison.Syntax.Printer.Type as TP
 import qualified Unison.UnisonFile as UF
 import qualified Unison.Util.Pretty as P
 import Unison.Var (Var)
