+++ unison-cli/src/Unison/CommandLine/OutputMessages.hs
@@ -81,26 +81,12 @@ import Unison.CommandLine.InputPatterns (makeExample')
 import qualified Unison.CommandLine.InputPatterns as IP
 import Unison.ConstructorReference (GConstructorReference (..))
 import qualified Unison.DataDeclaration as DD
-import qualified Unison.DeclPrinter as DeclPrinter
 import qualified Unison.Hash as Hash
 import qualified Unison.HashQualified as HQ
 import qualified Unison.HashQualified' as HQ'
 import Unison.LabeledDependency as LD
 import Unison.Name (Name)
 import qualified Unison.Name as Name
-import Unison.NamePrinter
-  ( prettyHashQualified,
-    prettyHashQualified',
-    prettyLabeledDependency,
-    prettyName,
-    prettyNamedReference,
-    prettyNamedReferent,
-    prettyReference,
-    prettyReferent,
-    prettyShortHash,
-    styleHashQualified,
-    styleHashQualified',
-  )
 import Unison.NameSegment (NameSegment (..))
 import Unison.Names (Names (..))
 import qualified Unison.Names as Names
@@ -133,11 +119,25 @@ import Unison.Share.Sync.Types (CodeserverTransportError (..))
 import qualified Unison.ShortHash as SH
 import qualified Unison.ShortHash as ShortHash
 import qualified Unison.Sync.Types as Share
+import qualified Unison.Syntax.Printer.Decl as DeclPrinter
+import Unison.Syntax.Printer.Name
+  ( prettyHashQualified,
+    prettyHashQualified',
+    prettyLabeledDependency,
+    prettyName,
+    prettyNamedReference,
+    prettyNamedReferent,
+    prettyReference,
+    prettyReferent,
+    prettyShortHash,
+    styleHashQualified,
+    styleHashQualified',
+  )
+import qualified Unison.Syntax.Printer.Term as TermPrinter
+import qualified Unison.Syntax.Printer.Type as TypePrinter
 import Unison.Term (Term)
 import qualified Unison.Term as Term
-import qualified Unison.TermPrinter as TermPrinter
 import Unison.Type (Type)
-import qualified Unison.TypePrinter as TypePrinter
 import qualified Unison.UnisonFile as UF
 import qualified Unison.Util.List as List
 import Unison.Util.Monoid (intercalateMap)
