+++ unison-share-api/src/Unison/Server/Doc.hs
@@ -31,8 +31,6 @@ import Unison.Codebase.Editor.DisplayObject (DisplayObject)
 import qualified Unison.Codebase.Editor.DisplayObject as DO
 import qualified Unison.ConstructorReference as ConstructorReference
 import qualified Unison.DataDeclaration as DD
-import qualified Unison.DeclPrinter as DeclPrinter
-import qualified Unison.NamePrinter as NP
 import qualified Unison.PrettyPrintEnv as PPE
 import qualified Unison.PrettyPrintEnvDecl as PPE
 import Unison.Reference (Reference)
@@ -44,12 +42,14 @@ import Unison.Server.Orphans ()
 import Unison.Server.Syntax (SyntaxText)
 import qualified Unison.Server.Syntax as Syntax
 import qualified Unison.ShortHash as SH
+import qualified Unison.Syntax.Printer.Decl as DeclPrinter
+import qualified Unison.Syntax.Printer.Name as NP
+import qualified Unison.Syntax.Printer.Term as TermPrinter
+import qualified Unison.Syntax.Printer.Type as TypePrinter
 import Unison.Term (Term)
 import qualified Unison.Term as Term
-import qualified Unison.TermPrinter as TermPrinter
 import Unison.Type (Type)
 import qualified Unison.Type as Type
-import qualified Unison.TypePrinter as TypePrinter
 import qualified Unison.Util.List as List
 import qualified Unison.Util.Pretty as P
 import qualified Unison.Util.SyntaxText as S
